pyproject.toml
[project]
authors = [
    { name = "Misael Gonzalez (Rayference)" },
    { name = "Nicolae Marton (Rayference)" },
]
description = "Observation simulator for the ESA DTE-S2GOS synthetic scene generator service."
name = "s2gos-simulator"
requires-python = ">=3.10"
version = "0.0.1"
dependencies = [
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
eradiate = [
    "eradiate[kernel]",
]
test = ["pytest"]
dev = []
docs = ["mkdocs"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
s2gos_simulator = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev", "test", "docs"], solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
ignore = ["E501", "F821"]
isort = { relative-imports-order = "closest-to-furthest" }
per-file-ignores = { "*.ipynb" = ["E402"] }
select = ["I", "F", "E"]

